insurance_assistant_system/requirements.txt
agno
openai
groq
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
pandas
tantivy
yfinance
psycopg2
pymilvus
google-generativeai
google-genai
reportlab
chonkie
sqlalchemy
fastapi
uvicorn
flask
streamlit-oauth
google-auth
chromadb